# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Everruns API",
    "description": "API for managing AI agents, sessions, messages, and events",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.2.0"
  },
  "servers": [
    {
      "url": "https://app.everruns.com/api",
      "description": "Production API"
    }
  ],
  "paths": {
...
    "/v1/orgs/{org}": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "summary": "GET /v1/orgs/:org - Get organization details",
        "operationId": "get_organization",
        "parameters": [
          {
            "name": "org",
            "in": "path",
            "description": "Organization public ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Organization details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationResponse"
                }
              }
            }
          },
          "404": {
            "description": "Organization not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Organizations"
        ],
        "summary": "PATCH /v1/orgs/:org - Update organization",
        "operationId": "update_organization",
        "parameters": [
          {
            "name": "org",
            "in": "path",
            "description": "Organization public ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganizationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Organization updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationResponse"
                }
              }
            }
          },
          "404": {
            "description": "Organization not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error response for API endpoints.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message describing what went wrong."
          }
        }
      },
...
      "OrganizationResponse": {
        "type": "object",
        "description": "Response for organization operations",
        "required": [
          "id",
          "name",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the organization was created"
          },
          "id": {
            "type": "string",
            "description": "External identifier (org_<32-hex-chars>)"
          },
          "name": {
            "type": "string",
            "description": "Display name"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the organization was last updated"
          }
        }
      },
...
      "UpdateOrganizationRequest": {
        "type": "object",
        "description": "Request to update an organization",
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ],
            "description": "The display name of the organization.",
            "example": "Acme Corporation"
          }
        }
      },
...
    }
  },
...
}